Update README.md
I put the code into my Titanium and did not find any errors. I thought your use of comments in the code was good and can be very useful. The Phone gadget that was used was the map gadget and the code was right and had no errors. This code was found in Chapter 7 of the book on page 173. One suggestion that I have is to change the colors that you used to what you want the theme to be. You used the same colors that were given in the book, and I think it would be better for the app if you designed a theme of colors to what you wanted them to be instead of using the ones that were provided.